# Improve performance of array integer indexing