This PR tries to improve performance of array integer indexing as a first step towards addressing #2878.

Since it seems like using `jit` as suggested in #2891 is difficult I've limited this PR to the case of indexing using a single Python integer outside of `jit` to keep the diff somewhat manageable. Some of the changes should also benefit other cases.

I split the PR into multiple commits that are best reviewed one by one. Also please let me know if these changes would prevent future planned optimisation that I am not aware of. If so I'm happy to close this, or split out some of the optimisation into separate PR's.

## Changes
- 09134b36bd0a3a6fc10609b8a7a28d93c69b2e93 prevents the conversion of Python integers to DeviceArrays and uses plain Python code in favour of `lax` operations for index normalisation when possible. This moves the scalar index computations to the CPU instead of running on an accelerator. I currently can't easily measure the performance impact of this but for me this seems to be preferred for statically known indices.
- 75a092e027b26bcecf832c8cbfe088e98ec8c08d prevents unnecessary calls to `lax.expand_dims`
- b8d70d31a5187c3c4cf72743b3f96a9026627904 removes explicit conversion using `asarray`
- cf8e8ed46062fce095ce76b001626a4d3c248a3f and 19ac51e10339c465f7c40ea3b9acb966320c07a0 remove expensive calls to `core.get_aval` to speedup checks

## Latency impact

I've measured indexing performance using a simple toy example on CPU:
```python
from jax import numpy as jnp

a = jnp.arange(4000)
a[5]
%timeit a[6].block_until_ready()
```

This PR improves performance by a factor of **4.8x**. Jax is still a lot slower than plain numpy but this already seems like a decent improvement.

| commit                                   |   time | reduction |
| ---------------------------------------- | -----: | --------: |
| 88ac4edf63625db10f92d5368a598205aac1deb4 (main baseline) | 468 µs |           |
| 09134b36bd0a3a6fc10609b8a7a28d93c69b2e93 | 225 µs | - 52 %    |
| 75a092e027b26bcecf832c8cbfe088e98ec8c08d | 168 µs | - 25 %    |
| b8d70d31a5187c3c4cf72743b3f96a9026627904 | 143 µs | - 15 %    |
| cf8e8ed46062fce095ce76b001626a4d3c248a3f | 125 µs | - 13 %    |
| 19ac51e10339c465f7c40ea3b9acb966320c07a0 |  98 µs | - 22 %    |