@lgeiger 2022-04-21
> Related: #10150

@jakevdp Thanks for mentioning that PR. I rebased and added a small test in 8054f426e13346d8c09022704761824651ef983c to ensure that only a gather operation is present when using constant integer indices.