@jakevdp 2022-04-25
It looks like this breaks some jax2tf tests, namely `jax/experimental/jax2tf/tests/shape_poly_test.py:ShapePolyPrimitivesTest.test_prim_getitem_op=poly_idx=slice-poly_noxla_poly_axes=[0]` (this test is not run as part of the github CI)