@jakevdp 2022-05-06
Thanks for the updates.

So, overall my concern here is that this PR touches a *lot* of parts of the code, all of which have potential side-effects that are difficult for me to predict. The passing tests are a good sign, but I'm not confident that our test coverage is complete when it comes to things like all the various ways that arrays can be indexed. A PR like this seems ripe for unanticipated and unintended downstream consequences.

Unfortunately, I'm not really sure what to suggest – it's hard to check for unknown issues. I'll pull-in the changes and run a larger suite of internal tests in case that shows us something.