@lgeiger 2022-05-09
Thanks for the detailed review and the help!
In case any unforeseen problems arise, I'm happy to take another look or extract  10addfd040acbbd1972a6697bf9c8134a1248043 and 38cee9f68644cb27307513ad56f2dad9134b5259 into separate PRs.